# product/activation-server/api/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "U-Claw Activation API",
    "version": "1.0.0"
  },
  "servers": [{ "url": "https://api.u-claw.org" }],
  "paths": {
...
    "/v1/client-policy": {
      "get": {
        "operationId": "getClientPolicy",
        "responses": {
          "200": {
            "description": "Current client policy",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ClientPolicy" } } }
          },
          "503": { "$ref": "#/components/responses/503" }
        }
      }
    },
    "/v1/activations": {
      "post": {
        "operationId": "createActivation",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ActivationRequest" } } }
        },
        "responses": {
          "200": {
            "description": "Activation bound to device",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ActivationResponse" } } }
          },
          "400": { "$ref": "#/components/responses/ActivationError" },
          "409": { "$ref": "#/components/responses/ActivationError" },
          "429": { "$ref": "#/components/responses/429" },
          "503": { "$ref": "#/components/responses/ActivationError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": { "type": "http", "scheme": "bearer" },
      "AdminBearerAuth": { "type": "http", "scheme": "bearer", "description": "Independent admin secret" }
    },
...
    "responses": {
      "ActivationError": {
        "description": "Stable redacted public error",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ActivationError" } } }
      },
...
      "429": {
        "description": "Rate limit exceeded",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ActivationError" } } }
      },
      "503": {
        "description": "Service unavailable",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ActivationError" } } }
      },
...
    },
    "schemas": {
      "Identifier": {
        "type": "string",
        "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{2,127}$"
      },
      "ActivationCode": { "type": "string", "pattern": "^[0-9A-HJKMNP-TV-Z]{26}$" },
...
      "UsbFingerprint": {
        "type": "object",
        "additionalProperties": false,
        "required": ["version", "sha256"],
        "properties": {
          "version": { "const": "uclaw-usb-v1" },
          "sha256": { "type": "string", "pattern": "^[a-f0-9]{64}$" }
        }
      },
      "ActivationRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["activationCode", "usbFingerprint", "clientVersion", "idempotencyKey"],
        "properties": {
          "activationCode": { "$ref": "#/components/schemas/ActivationCode" },
          "usbFingerprint": { "$ref": "#/components/schemas/UsbFingerprint" },
          "clientVersion": { "type": "string", "pattern": "^(?:0|[1-9][0-9]*)\\.(?:0|[1-9][0-9]*)\\.(?:0|[1-9][0-9]*)$" },
          "idempotencyKey": { "type": "string", "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{7,127}$" }
        }
      },
      "ActivationResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["activationId", "deviceId", "licenseId", "license", "startupCredential", "builtinCredential", "status"],
        "properties": {
          "activationId": { "$ref": "#/components/schemas/Identifier" },
          "deviceId": { "$ref": "#/components/schemas/Identifier" },
          "licenseId": { "$ref": "#/components/schemas/Identifier" },
          "license": { "$ref": "#/components/schemas/StartupLicense" },
          "startupCredential": { "$ref": "#/components/schemas/StartupCredential" },
          "builtinCredential": { "$ref": "#/components/schemas/BuiltinCredential" },
          "status": { "const": "active" }
        }
      },
...
      "ActivationError": {
        "type": "object",
        "additionalProperties": false,
        "required": ["requestId", "activationId", "code", "stage", "retryable", "supportCode"],
        "properties": {
          "requestId": { "$ref": "#/components/schemas/Identifier" },
          "activationId": { "oneOf": [{ "$ref": "#/components/schemas/Identifier" }, { "type": "null" }] },
          "code": { "type": "string", "pattern": "^[A-Z][A-Z0-9_]{2,63}$" },
          "stage": {
            "type": ["string", "null"],
            "enum": ["requested", "server_bound", "committed", "failed_before_bind", null]
          },
          "retryable": { "type": "boolean" },
          "supportCode": { "type": "string", "pattern": "^[A-Z0-9][A-Z0-9-]{2,31}$" }
        }
      },
      "ClientPolicy": {
        "type": "object",
        "additionalProperties": false,
        "description": "Signed short-lived global production policy. Client startup continues only when localReleaseSequence exactly equals requiredReleaseSequence.",
        "required": ["schemaVersion", "policyEpoch", "requiredReleaseSequence", "releaseId", "contentVersion", "reason", "manifestUrl", "manifestSha256", "issuedAt", "expiresAt", "signature"],
        "properties": {
          "schemaVersion": { "const": 1 },
          "policyEpoch": { "type": "integer", "minimum": 1, "maximum": 9007199254740991 },
          "requiredReleaseSequence": { "type": "integer", "minimum": 1, "maximum": 9007199254740991 },
          "releaseId": { "$ref": "#/components/schemas/ReleaseId" },
          "contentVersion": { "type": "string", "pattern": "^[A-Za-z0-9][A-Za-z0-9.+_-]{0,63}$" },
          "reason": { "enum": ["release", "rollback"] },
          "manifestUrl": { "type": "string", "format": "uri", "pattern": "^https://" },
          "manifestSha256": { "$ref": "#/components/schemas/SHA256" },
          "issuedAt": { "type": "string", "format": "date-time" },
          "expiresAt": { "type": "string", "format": "date-time" },
          "signature": { "$ref": "#/components/schemas/ReleasePolicySignature" }
        }
      },
      "ReleaseId": { "type": "string", "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,127}$" },
      "SHA256": { "type": "string", "pattern": "^[a-f0-9]{64}$" },
      "ReleasePolicySignature": {
        "type": "object", "additionalProperties": false, "required": ["algorithm", "keyId", "value"],
        "properties": { "algorithm": { "const": "ed25519" }, "keyId": { "$ref": "#/components/schemas/ReleaseId" }, "value": { "type": "string", "minLength": 88, "maxLength": 88 } }
      },
...
      "StartupLicense": {
        "type": "object",
        "additionalProperties": false,
        "required": ["schemaVersion", "usernameId", "deviceId", "licenseId", "usbFingerprint", "startupSecretProof", "notBefore", "expiresAt", "revision", "signature"],
        "properties": {
          "schemaVersion": { "const": 1 },
          "usernameId": { "$ref": "#/components/schemas/Identifier" },
          "deviceId": { "$ref": "#/components/schemas/Identifier" },
          "licenseId": { "$ref": "#/components/schemas/Identifier" },
          "usbFingerprint": {
            "type": "object", "additionalProperties": false, "required": ["scheme", "sha256"],
            "properties": { "scheme": { "const": "uclaw-usb-v1" }, "sha256": { "type": "string", "pattern": "^[a-f0-9]{64}$" } }
          },
          "startupSecretProof": {
            "type": "object", "additionalProperties": false,
            "required": ["algorithm", "startupSecretSalt", "startupSecretHash"],
            "properties": {
              "algorithm": { "const": "sha256-salt-v1" },
              "startupSecretSalt": { "type": "string", "pattern": "^(?:[a-f0-9]{2}){16,64}$" },
              "startupSecretHash": { "type": "string", "pattern": "^[a-f0-9]{64}$" }
            }
          },
          "notBefore": { "type": "string", "format": "date-time", "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}Z$" },
          "expiresAt": { "type": "string", "format": "date-time", "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}Z$" },
          "revision": { "type": "integer", "minimum": 1, "maximum": 9007199254740991 },
          "signature": {
            "type": "object", "additionalProperties": false, "required": ["algorithm", "keyId", "value"],
            "properties": {
              "algorithm": { "const": "ed25519" },
              "keyId": { "$ref": "#/components/schemas/Identifier" },
              "value": { "type": "string", "minLength": 80, "maxLength": 256 }
            }
          }
        }
      },
      "StartupCredential": {
        "type": "object",
        "additionalProperties": false,
        "required": ["schemaVersion", "deviceId", "licenseId", "startupSecret"],
        "properties": {
          "schemaVersion": { "const": 1 },
          "deviceId": { "$ref": "#/components/schemas/Identifier" },
          "licenseId": { "$ref": "#/components/schemas/Identifier" },
          "startupSecret": { "type": "string", "minLength": 32, "maxLength": 512 }
        }
      },
      "BuiltinCredential": {
        "type": "object",
        "additionalProperties": false,
        "required": ["schemaVersion", "deviceId", "licenseId", "endpoint", "deviceToken"],
        "properties": {
          "schemaVersion": { "const": 2 },
          "deviceId": { "$ref": "#/components/schemas/Identifier" },
          "licenseId": { "$ref": "#/components/schemas/Identifier" },
          "endpoint": { "type": "string", "format": "uri", "pattern": "^[Hh][Tt][Tt][Pp][Ss]://(?:\\[[0-9A-Fa-f:.]+\\]|[^/?#@\\[\\]\\\\\\u0000-\\u0020\\u007F:]+)(?::[0-9]+)?(?:/[^?#@\\\\\\u0000-\\u0020\\u007F]*)?$" },
          "deviceToken": { "type": "string", "pattern": "^uclaw_dt_[A-Za-z0-9_-]{43}$" }
        }
      },
...
    }
  }
}